#279 cleanup and refactor
