+++ components/FacilityDetails/table-cols.js
@@ -1,0 +1,43 @@
+const tableCols = [
+  {
+    title: 'First Name',
+    field: 'first_name',
+  },
+  {
+    title: 'Last Name',
+    field: 'last_name',
+  },
+  {
+    title: 'Email',
+    field: 'email',
+  },
+  {
+    title: 'Gender',
+    field: 'gender',
+  },
+  {
+    title: 'Phone',
+    field: 'phone',
+  },
+  {
+    title: 'Address',
+    field: 'address',
+  },
+
+  {
+    title: 'City',
+    field: 'city',
+  },
+
+  {
+    title: 'State',
+    field: 'state',
+  },
+
+  {
+    title: 'Zip',
+    field: 'zip',
+  },
+]
+
+export default tableCols
