+++ components/FacilityDetails/use-styles.js
@@ -1,0 +1,16 @@
+import { makeStyles } from '@material-ui/core/styles'
+
+const useStyles = makeStyles((theme) => ({
+  h2: {
+    marginTop: '.5em',
+  },
+  tableHead: {
+    fontWeight: 'bold',
+  },
+  error: {
+    color: 'red',
+    fontWeight: 'bold',
+  },
+}))
+
+export default useStyles
