+++ pages/facility/facility-details/[facility_uuid].js
@@ -1,56 +1,13 @@
-import React, { useContext, useEffect, useState } from 'react'
+import React, { useEffect, useState } from 'react'
 import { NextSeo } from 'next-seo'
-import xhrHeader from '../../../constants/xhrHeader'
-import {
-  Typography,
-  Box,
-  Table,
-  TableRow,
-  TableCell,
-  CircularProgress,
-} from '@material-ui/core'
-import Container from '../../../components/Container'
 import { makeStyles } from '@material-ui/core/styles'
 import { useRouter } from 'next/router'
-import MaterialTable from 'material-table'
-import { Skeleton } from '@material-ui/lab'
-const useStyles = makeStyles((theme) => ({
-  h2: {
-    marginTop: '.5em',
-  },
-  textFields: {
-    width: '100%',
-    marginTop: '2em',
-    maxWidth: '34rem',
-  },
-  tableHead: {
-    fontWeight: 'bold',
-  },
-  error: {
-    color: 'red',
-    fontWeight: 'bold',
-  },
-  buttonLinks: {
-    '@media screen and (max-width: 700px)': {
-      '&:nth-child(2)': {
-        order: -1,
-      },
-    },
+import { Typography, Box, CircularProgress } from '@material-ui/core'
+import xhrHeader from '../../../constants/xhrHeader'
+import Container from '../../../components/Container'
+import FacilityDetails from '../../../components/FacilityDetails'
 
-    '& button': {
-      padding: '1em',
-      fontWeight: 600,
-      width: '16rem',
-
-      '&:hover': {
-        backgroundColor: theme.palette.primary.main,
-      },
-    },
-  },
-}))
-
-const FacilityDetails = () => {
-  const classes = useStyles()
+const FacilityDetailsPage = () => {
   const [facility, setFacility] = useState()
   const [error, setError] = useState(false)
   const [loading, setLoading] = useState(true)
@@ -110,118 +67,7 @@ const FacilityDetails = () => {
             </Box>
           ) : (
             <>
-              <Typography variant="h2" className={classes.h2}>
-                {facility?.name}
-              </Typography>
-
-              <Typography variant="h3" className={classes.h2}>
-                Facility Information
-              </Typography>
-              <Table>
-                <TableRow>
-                  <TableCell variant="head" className={classes.tableHead}>
-                    Address:
-                  </TableCell>
-                  <TableCell>
-                    {facility?.address}
-                    <br />
-                    {facility?.city}, {facility?.state} {facility?.zip}
-                  </TableCell>
-                </TableRow>
-                <TableRow>
-                  <TableCell variant="head" className={classes.tableHead}>
-                    Primary Contact Name:
-                  </TableCell>
-                  <TableCell>{facility?.primary_contact_name}</TableCell>
-                </TableRow>
-                <TableRow>
-                  <TableCell variant="head" className={classes.tableHead}>
-                    Primary Contact Phone:
-                  </TableCell>
-                  <TableCell>
-                    {facility?.primary_contact_mobile_phone}
-                  </TableCell>
-                </TableRow>
-                <TableRow>
-                  <TableCell variant="head" className={classes.tableHead}>
-                    Primary Contact Shift:
-                  </TableCell>
-                  <TableCell>{facility?.primary_contact_shift}</TableCell>
-                </TableRow>
-                {facility?.secondary_contact_name && (
-                  <TableRow>
-                    <TableCell variant="head" className={classes.tableHead}>
-                      Secondary Contact Name:
-                    </TableCell>
-                    <TableCell>{facility?.secondary_contact_name}</TableCell>
-                  </TableRow>
-                )}
-
-                {facility?.secondary_contact_mobile_phone && (
-                  <TableRow>
-                    <TableCell variant="head" className={classes.tableHead}>
-                      Secondary Contact Phone:
-                    </TableCell>
-                    <TableCell>
-                      {facility?.secondary_contact_mobile_phone}
-                    </TableCell>
-                  </TableRow>
-                )}
-
-                {facility?.secondary_contact_shift && (
-                  <TableRow>
-                    <TableCell variant="head" className={classes.tableHead}>
-                      Secondary Contact Shift:
-                    </TableCell>
-                    <TableCell>{facility?.secondary_contact_shift}</TableCell>
-                  </TableRow>
-                )}
-              </Table>
-              <MaterialTable
-                title="Patient Information"
-                columns={[
-                  {
-                    title: 'First Name',
-                    field: 'first_name',
-                  },
-                  {
-                    title: 'Last Name',
-                    field: 'last_name',
-                  },
-                  {
-                    title: 'Email',
-                    field: 'email',
-                  },
-                  {
-                    title: 'Gender',
-                    field: 'gender',
-                  },
-                  {
-                    title: 'Phone',
-                    field: 'phone',
-                  },
-                  {
-                    title: 'Address',
-                    field: 'address',
-                  },
-
-                  {
-                    title: 'City',
-                    field: 'city',
-                  },
-
-                  {
-                    title: 'State',
-                    field: 'state',
-                  },
-
-                  {
-                    title: 'Zip',
-                    field: 'zip',
-                  },
-                ]}
-                data={facility?.patients}
-              />
+              <FacilityDetails facility={facility} />
             </>
           )}
         </Container>
@@ -230,4 +76,4 @@ const FacilityDetails = () => {
   )
 }
 
-export default FacilityDetails
+export default FacilityDetailsPage
